+++ packages/selector-resolve-nested/test/index.mjs
@@ -3,3 +3,4 @@ import './compound.mjs';
 import './lists.mjs';
 import './simple-compound-complex.mjs';
 import './multiple-resolves.mjs';
+import './source-indices.mjs';