+++ packages/selector-resolve-nested/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to Selector Resolve Nested
 
+### Unreleased (patch)
+
+- Improve source index resolution
+
 ### 1.0.0
 
 _January 21, 2024_